# backend/docs/swagger.yaml
components:
  schemas:
...
    data:
      properties:
        data:
          properties:
            message:
              type: string
          type: object
      type: object
...
    relationships.RelationshipRequest:
      properties:
        description:
          type: string
        relationship_type:
          type: string
        source_entity_id:
          type: integer
        source_entity_type:
          type: string
        target_entity_id:
          type: integer
        target_entity_type:
          type: string
        work_id:
          type: integer
      required:
      - relationship_type
      - source_entity_id
      - source_entity_type
      - target_entity_id
      - target_entity_type
      type: object
...
    response.StandardResponse:
      allOf:
      - $ref: '#/components/schemas/data'
      properties:
        code:
          type: integer
        data: {}
        message:
          type: string
        sourceId:
          type: string
        traceId:
          type: string
      type: object
...
  securitySchemes:
    BearerAuth:
      in: header
      name: Authorization
      type: apiKey
...
info:
  description: This is the API for Test.
  title: API 1.0
  version: "1.0"
openapi: 3.1.0
paths:
...
  /relationships:
    get:
      description: List relationships with pagination and filters
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        schema:
          type: integer
      - description: 'Number of items per page (default: 10)'
        in: query
        name: limit
        schema:
          type: integer
      - description: Filter by Source Entity ID
        in: query
        name: sourceEntityId
        schema:
          type: integer
      - description: Filter by Source Entity Type
        in: query
        name: sourceEntityType
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/data'
                properties:
                  code:
                    type: integer
                  data: {}
                  message:
                    type: string
                  sourceId:
                    type: string
                  traceId:
                    type: string
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Invalid filter parameters
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Failed to retrieve relationships
      security:
      - BearerAuth: []
      summary: List relationships
      tags:
      - relationships
    post:
      description: Create a new relationship between entities
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/relationships.RelationshipRequest'
        description: Relationship info
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/data'
                properties:
                  code:
                    type: integer
                  data: {}
                  message:
                    type: string
                  sourceId:
                    type: string
                  traceId:
                    type: string
                type: object
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Invalid request body
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Failed to create relationship
      security:
      - BearerAuth: []
      summary: Create a new relationship
      tags:
      - relationships
  /relationships/{id}:
    delete:
      description: Delete a relationship by its ID
      parameters:
      - description: Relationship ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/data'
                properties:
                  code:
                    type: integer
                  data: {}
                  message:
                    type: string
                  sourceId:
                    type: string
                  traceId:
                    type: string
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Invalid ID
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Relationship not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Failed to delete relationship
      security:
      - BearerAuth: []
      summary: Delete a relationship
      tags:
      - relationships
    get:
      description: Get a relationship by its ID
      parameters:
      - description: Relationship ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/data'
                properties:
                  code:
                    type: integer
                  data: {}
                  message:
                    type: string
                  sourceId:
                    type: string
                  traceId:
                    type: string
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Invalid ID
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Relationship not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Failed to get relationship
      security:
      - BearerAuth: []
      summary: Get a relationship by ID
      tags:
      - relationships
    put:
      description: Update a relationship by its ID
      parameters:
      - description: Relationship ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/relationships.RelationshipRequest'
        description: Relationship info
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/data'
                properties:
                  code:
                    type: integer
                  data: {}
                  message:
                    type: string
                  sourceId:
                    type: string
                  traceId:
                    type: string
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Invalid ID or request body
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Relationship not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Failed to update relationship
      security:
      - BearerAuth: []
      summary: Update a relationship
      tags:
      - relationships
...